create verb optin for cli
